services/api/requirements.txt
#server/requests
httpx==0.28.1
fastapi[standard]==0.115.7
uvicorn==0.34.0
httptools==0.6.4

#utils
loguru==0.7.3
tzlocal==5.3
pytz==2025.1

#db/orm/redis
sqlalchemy==2.0.38
#asyncpg==0.30.0
aiomysql==0.2.0
sqlmodel==0.0.24
redis==5.2.1

#security
pyjwt==2.10.1
passlib[bcrypt]==1.7.4
cryptography==44.0.2

